remove form, contract.role and load veevalidate via localvue
