--- src/components/UWInput/index.js
@@ -1,4 +1,0 @@
-import ContactForm from './ContactForm.vue'
-import UWInput from './UWInput.vue'
-
-export { ContactForm, UWInput }
