+++ src/components/index.js
@@ -1,7 +1,9 @@
 import Version from 'src/components/Version.js'
+import UWInput from 'src/components/UWInput'
 
 export * from 'src/components/ModalWrapperElements'
 
 export {
   Version,
+  UWInput,
 }
