--- tests/helpers/school-contacts.ts
@@ -1,4 +1,0 @@
-
-export const ContactRoles: object = {
-  CONTACT: <string> 'default-contact-role',
-}
